refactor: move rpg-style to data folder
