+++ src/infra/db/prisma/data/rpg-style/add-rpg-style/add-rpg-style-prisma-repo.spec.ts
@@ -1,7 +1,7 @@
 import { type RpgStyleModel } from '@/domain/models'
 import { AddRpgStylePrismaRepo } from './add-rpg-style-prisma-repo'
 import { PrismockClient } from 'prismock'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 import { type PrismaClient } from '@prisma/client'
 
 let prismock: PrismaClient
