+++ src/infra/db/prisma/data/rpg-style/add-rpg-style/add-rpg-style-prisma-repo.ts
@@ -1,6 +1,6 @@
 import { type RpgStyleModel } from '@/domain/models'
 import { type AddRpgStyleRepo } from '@/usecases/contracts/db/rpg-style'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 export class AddRpgStylePrismaRepo implements AddRpgStyleRepo {
   async execute (data: RpgStyleModel): Promise<void> {
