+++ src/infra/db/prisma/data/rpg-style/find-rpg-style-by-name/find-rpg-style-by-name-prisma-repo.spec.ts
@@ -2,7 +2,7 @@ import { type RpgStyleModel } from '@/domain/models'
 import { type PrismaClient } from '@prisma/client'
 import { FindRpgStyleByNamePrismaRepo } from './find-rpg-style-by-name-prisma-repo'
 import { PrismockClient } from 'prismock'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 let prismock: PrismaClient
