+++ src/infra/db/prisma/data/rpg-style/find-rpg-style-by-name/find-rpg-style-by-name-prisma-repo.ts
@@ -1,5 +1,5 @@
 import { type FindRpgStyleByNameRepo } from '@/usecases/contracts/db/rpg-style'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 import { type RpgStyleModel } from '@/domain/models'
 
 export class FindRpgStyleByNamePrismaRepo implements FindRpgStyleByNameRepo {
