+++ src/main/factories/infra/db/prisma/rpg-style/add-rpg-style-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddRpgStylePrismaRepo } from '@/infra/db/prisma/rpg-style/add-rpg-style/add-rpg-style-prisma-repo'
+import { AddRpgStylePrismaRepo } from '@/infra/db/prisma/data/rpg-style/add-rpg-style/add-rpg-style-prisma-repo'
 
 export const makeAddRpgStylePrismaRepo = (): AddRpgStylePrismaRepo => {
   return new AddRpgStylePrismaRepo()
