+++ src/main/factories/infra/db/prisma/rpg-style/find-many-rpg-styles-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindManyRpgStylesPrismaRepo } from '@/infra/db/prisma/rpg-style/find-many-rpg-styles/find-many-rpg-styles-prisma-repo'
+import { FindManyRpgStylesPrismaRepo } from '@/infra/db/prisma/data/rpg-style/find-many-rpg-styles/find-many-rpg-styles-prisma-repo'
 
 export const makeFindManyRpgStylesPrismaRepo = (): FindManyRpgStylesPrismaRepo => {
   return new FindManyRpgStylesPrismaRepo()
