+++ src/main/factories/infra/db/prisma/rpg-style/find-rpg-style-by-name-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindRpgStyleByNamePrismaRepo } from '@/infra/db/prisma/rpg-style/find-rpg-style-by-name/find-rpg-style-by-name-prisma-repo'
+import { FindRpgStyleByNamePrismaRepo } from '@/infra/db/prisma/data/rpg-style/find-rpg-style-by-name/find-rpg-style-by-name-prisma-repo'
 
 export const makeFindRpgStyleByNamePrismaRepo = (): FindRpgStyleByNamePrismaRepo => {
   return new FindRpgStyleByNamePrismaRepo()
